Separate all screens into their own files. Added querying from the database to select songs. Search feature supports, artist, song title, and genre
